+++ docs/changelog.rst
@@ -6,6 +6,9 @@ All notable changes to this project will be documented in this page.
 The format is based on `Keep a Changelog`_, and this project adheres to
 `Semantic Versioning`_.
 
+Unreleased
+==========
+
 v0.2.1 (22 November 2022)
 =========================
 New setup_staticmaps_from_raster method and river smoothing algorithm. Correct some bugs linked to soon